fix: ensure compact_tx format is respected
Right now, the values implementation is quite a bit more bloated than
it needs to be; the input side of a transaction only needs to be the
UHS ids (just the hashes), not the full UHS element.

Because the benchmarks were largely written against the cryptographic
commitment code (for which this optimization has an outsized effect),
they assumed this would be the case.

For now, instead of implementing this optimization in the values
solution, we just ensure the benchmarks behave correctly.

In the long-run, if this branch should stay around, we should
correctly minimize the input-side of compact transactions.

Signed-off-by: Sam Stuewe <stuewe@mit.edu>