feat: verify rangeproofs in sentinel
This also disables the rangeproof verification at the shard-layer. So
long as the sentinels are inside the trusted perimeter, or the number
of compromised sentinels remains lower than the specified attestation
threshold (as verified by the coordinators), this has no drawback.

Furthermore, it likely offers significant performance benefits by
allowing the most horizontally-scalable component to handle the
heaviest verification. Moreover, by checking the proof at this point
in the flow, the range is vetted before the transaction is committed
to the UHS.

Signed-off-by: Sam Stuewe <stuewe@mit.edu>